pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scaffold"
version = "0.1.0"
description = "Deploy any service stack to Railway/Vercel — zero-touch, agent-friendly"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "anthropic>=0.40",
    "python-dotenv>=1.0",
]

[project.scripts]
scaffold = "scaffold.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["scaffold"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
